requirements.txt
openai
pymongo
telegram
python-telegram-bot
langchain
python-dotenv
pydantic
pydub